+++ packages/moengage_cards/moengage_cards_android/android/src/main/kotlin/com/moengage/flutter/cards/EventEmitterImpl.kt
@@ -1,12 +1,12 @@
 package com.moengage.flutter.cards
 
-import com.moengage.core.LogLevel
-import com.moengage.core.internal.logger.Logger
 import com.moengage.plugin.base.cards.internal.CardsEventEmitter
 import com.moengage.plugin.base.cards.internal.cardsSyncToJson
 import com.moengage.plugin.base.cards.internal.model.events.CardEventType
 import com.moengage.plugin.base.cards.internal.model.events.CardsEvent
 import com.moengage.plugin.base.cards.internal.model.events.CardsSyncEvent
+import com.moengage.platform.internal.logger.Logger
+import com.moengage.platform.internal.logger.PlatformLogLevel
 import org.json.JSONObject
 
 class EventEmitterImpl(private val callBack: (methodName: String, payload: String) -> Unit) :
@@ -17,10 +17,10 @@ class EventEmitterImpl(private val callBack: (methodName: String, payload: Strin
         try {
             when (event) {
                 is CardsSyncEvent -> emitCardSyncEvent(event)
-                else -> Logger.print(LogLevel.ERROR) { "$tag emit() Unknown Event: $event" }
+                else -> Logger.record(PlatformLogLevel.ERROR) { "$tag emit() Unknown Event: $event" }
             }
         } catch (t: Throwable) {
-            Logger.print(LogLevel.ERROR, t) { "emit(): $event" }
+            Logger.record(PlatformLogLevel.ERROR, t) { "emit(): $event" }
         }
     }
 
@@ -28,7 +28,7 @@ class EventEmitterImpl(private val callBack: (methodName: String, payload: Strin
         try {
             val syncCompleteData = event.syncCompleteData
             if (syncCompleteData == null) {
-                Logger.print(LogLevel.ERROR) { "emitCardSyncEvent(): $event : Sync CompleteData is null" }
+                Logger.record(PlatformLogLevel.ERROR) { "emitCardSyncEvent(): $event : Sync CompleteData is null" }
             }
             val syncCompleteJson = cardsSyncToJson(syncCompleteData, event.accountMeta)
             val method =
@@ -39,7 +39,7 @@ class EventEmitterImpl(private val callBack: (methodName: String, payload: Strin
                 }
             emit(method, syncCompleteJson)
         } catch (t: Throwable) {
-            Logger.print(LogLevel.ERROR, t) { "$tag emitCardSyncEvent(): $event" }
+            Logger.record(PlatformLogLevel.ERROR, t) { "$tag emitCardSyncEvent(): $event" }
         }
     }
 
@@ -48,10 +48,10 @@ class EventEmitterImpl(private val callBack: (methodName: String, payload: Strin
         payload: JSONObject,
     ) {
         try {
-            Logger.print { "$tag emit() : methodName: $methodName , payload: $payload" }
+            Logger.record { "$tag emit() : methodName: $methodName , payload: $payload" }
             callBack.invoke(methodName, payload.toString())
         } catch (t: Throwable) {
-            Logger.print(LogLevel.ERROR, t) { "$tag emit() : " }
+            Logger.record(PlatformLogLevel.ERROR, t) { "$tag emit() : " }
         }
     }
 }