+++ packages/moengage_personalize/moengage_personalize_android/android/src/main/kotlin/com/moengage/flutter/personalize/MoEngagePersonalizePlugin.kt
@@ -1,9 +1,9 @@
 package com.moengage.flutter.personalize
 
 import android.content.Context
-import com.moengage.core.LogLevel
-import com.moengage.core.internal.logger.Logger
 import com.moengage.plugin.base.personalization.PersonalizationHelper
+import com.moengage.platform.internal.logger.Logger
+import com.moengage.platform.internal.logger.PlatformLogLevel
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.embedding.engine.plugins.FlutterPlugin.FlutterPluginBinding
 import io.flutter.embedding.engine.plugins.activity.ActivityAware
@@ -20,19 +20,19 @@ class MoEngagePersonalizePlugin : FlutterPlugin, ActivityAware {
 
     override fun onAttachedToEngine(binding: FlutterPluginBinding) {
         try {
-            Logger.print { "$tag onAttachedToEngine() : Registering MoEngagePersonalizePlugin" }
+            Logger.record { "$tag onAttachedToEngine() : Registering MoEngagePersonalizePlugin" }
             context = binding.applicationContext
             flutterPluginBinding = binding
             if (methodChannel == null) {
                 initPlugin(binding.binaryMessenger)
             }
         } catch (t: Throwable) {
-            Logger.print(LogLevel.ERROR, t) { "$tag onAttachedToEngine()  : " }
+            Logger.record(PlatformLogLevel.ERROR, t) { "$tag onAttachedToEngine()  : " }
         }
     }
 
     private fun initPlugin(binaryMessenger: BinaryMessenger) {
-        Logger.print { "$tag initPlugin(): Initializing MoEngage Personalize Plugin" }
+        Logger.record { "$tag initPlugin(): Initializing MoEngage Personalize Plugin" }
         methodChannel = MethodChannel(binaryMessenger, CHANNEL_NAME)
         methodChannel?.setMethodCallHandler(
             PlatformMethodCallHandler(
@@ -43,7 +43,7 @@ class MoEngagePersonalizePlugin : FlutterPlugin, ActivityAware {
     }
 
     override fun onDetachedFromEngine(binding: FlutterPluginBinding) {
-        Logger.print { "$tag onDetachedFromEngine() : Detaching the Framework" }
+        Logger.record { "$tag onDetachedFromEngine() : Detaching the Framework" }
     }
 
     /**
@@ -51,7 +51,7 @@ class MoEngagePersonalizePlugin : FlutterPlugin, ActivityAware {
      * @param binding instance of [ActivityPluginBinding]
      */
     override fun onAttachedToActivity(binding: ActivityPluginBinding) {
-        Logger.print { "$tag onAttachedToActivity() : Attached To Activity" }
+        Logger.record { "$tag onAttachedToActivity() : Attached To Activity" }
         flutterPluginBinding?.binaryMessenger?.let {
             initPlugin(it)
         }
@@ -61,7 +61,7 @@ class MoEngagePersonalizePlugin : FlutterPlugin, ActivityAware {
      * Called when the plugin is Detached From Flutter Activity.
      */
     override fun onDetachedFromActivity() {
-        Logger.print { "$tag onDetachedFromActivity() : Resetting methodChannel to `null`" }
+        Logger.record { "$tag onDetachedFromActivity() : Resetting methodChannel to `null`" }
         methodChannel = null
     }
 
@@ -69,7 +69,7 @@ class MoEngagePersonalizePlugin : FlutterPlugin, ActivityAware {
      * Called when the plugin is Detached From Flutter Activity for Config Changes
      */
     override fun onDetachedFromActivityForConfigChanges() {
-        Logger.print {
+        Logger.record {
             "$tag onDetachedFromActivityForConfigChanges() : Detached From Activity for Config changes"
         }
     }
@@ -79,7 +79,7 @@ class MoEngagePersonalizePlugin : FlutterPlugin, ActivityAware {
      * @param binding instance of [ActivityPluginBinding]
      */
     override fun onReattachedToActivityForConfigChanges(binding: ActivityPluginBinding) {
-        Logger.print {
+        Logger.record {
             "$tag onReattachedToActivityForConfigChanges() : ReAttached To Activity for Config changes"
         }
     }