+++ packages/moengage_inbox/moengage_inbox_ios/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## Release Version
 
-- [major] Updated minimum supported Dart SDK to `3.6.0` and minimum Flutter SDK to `3.27.0`
+- [major] Updated minimum supported Dart SDK to `3.6.0` and minimum Flutter SDK to `3.44.0`
 
 # 13-08-2026